+++ Tests/Bootstrap_Test.lua
@@ -9,18 +9,34 @@
 ----------------------------------------------------------------------
 
 ----------------------------------------------------------------------
--- ACP Studio
--- Bootstrap_Test.lua
+-- Technical Bootstrap
+----------------------------------------------------------------------
+
+dofile(
+    debug.getinfo(1, "S").source:match("@?(.*[/\\])") ..
+    "TestSetup.lua"
+)
+----------------------------------------------------------------------
+-- Bootstrap
 ----------------------------------------------------------------------
 
 local Bootstrap = require("Tests.Bootstrap")
 
 assert(Bootstrap.Initialize())
+assert(Bootstrap.IsInitialized())
+assert(Bootstrap.GetRepositoryRoot() ~= nil)
+
+----------------------------------------------------------------------
+-- Repository Access
+----------------------------------------------------------------------
 
 local Logger = require("Core.Logger")
 
+assert(Logger ~= nil)
+
+----------------------------------------------------------------------
+-- Result
+----------------------------------------------------------------------
 
 reaper.ClearConsole()
 reaper.ShowConsoleMsg("Bootstrap_Test : PASS\n")
-
-